style: Move MetaMission to correct top level package
Ref #1551
