+++ mars-sim-core/src/main/java/com/mars_sim/core/mission/predefined/TestDriveMetaMission.java
@@ -8,11 +8,11 @@ package com.mars_sim.core.mission.predefined;
 
 import java.util.Set;
 
+import com.mars_sim.core.mission.AbstractMetaMission;
 import com.mars_sim.core.person.Person;
 import com.mars_sim.core.person.ai.job.util.JobType;
 import com.mars_sim.core.person.ai.mission.Mission;
 import com.mars_sim.core.person.ai.mission.MissionType;
-import com.mars_sim.core.person.ai.mission.meta.AbstractMetaMission;
 import com.mars_sim.core.vehicle.VehicleType;
 
 /**
