+++ mars-sim-core/src/main/java/com/mars_sim/core/mission/util/MissionRating.java
@@ -8,7 +8,7 @@ package com.mars_sim.core.mission.util;
 
 import com.mars_sim.core.data.RatedActivity;
 import com.mars_sim.core.data.RatingScore;
-import com.mars_sim.core.person.ai.mission.meta.MetaMission;
+import com.mars_sim.core.mission.MetaMission;
 
 /**
  * This represents a mission selection that has a meta mission and a rating score.
