+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/mission/meta/RescueSalvageVehicleMeta.java
@@ -10,6 +10,7 @@ import java.util.Collections;
 import java.util.Set;
 
 import com.mars_sim.core.data.RatingScore;
+import com.mars_sim.core.mission.AbstractMetaMission;
 import com.mars_sim.core.person.Person;
 import com.mars_sim.core.person.ai.job.util.JobType;
 import com.mars_sim.core.person.ai.mission.Mission;
